requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
jinja2
python-multipart
pypdf
pdfplumber